simplify how series are generated

+++ src/main/java/com/backpackcloud/sherlogholmes/domain/chart/Series.java
@@ -31,7 +31,7 @@ public interface Series {
 
   List<Bucket> buckets();
 
-  int total();
+  long total();
 
   Series add(String name, Series other);
